Move Revenants into their own plugin class instead of piggybacking the combat_script_binding
